+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/impl/DefaultSampleService.java
@@ -47,7 +47,6 @@ import uk.ac.bbsrc.tgac.miso.core.data.VolumeUnit;
 import uk.ac.bbsrc.tgac.miso.core.data.Workset;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleIdentityImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleIdentityImpl.IdentityBuilder;
-import uk.ac.bbsrc.tgac.miso.core.data.impl.SampleImpl;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.transfer.Transfer;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.transfer.TransferSample;
 import uk.ac.bbsrc.tgac.miso.core.exception.MisoNamingException;
@@ -243,7 +242,7 @@ public class DefaultSampleService implements SampleService, PaginatedDataSource<
         if (detailed.getParent() == null) {
           throw new IllegalArgumentException(ERR_MISSING_PARENT_ID);
         }
-        if (detailed.getParent().getId() != SampleImpl.UNSAVED_ID) {
+        if (detailed.getParent().isSaved()) {
           detailed.setParent((DetailedSample) get(detailed.getParent().getId()));
         } else {
           try {
@@ -352,7 +351,7 @@ public class DefaultSampleService implements SampleService, PaginatedDataSource<
       if (!hasTemporaryAlias(sample)) {
         validateAlias(sample, namingScheme);
       }
-      if (newId == Sample.UNSAVED_ID) {
+      if (!sample.isSaved()) {
         newId = sampleStore.addSample(sample);
       } else {
         sampleStore.update(sample);
@@ -521,7 +520,7 @@ public class DefaultSampleService implements SampleService, PaginatedDataSource<
       throw new IllegalArgumentException(ERR_MISSING_PARENT_ID);
     }
     DetailedSample tempParent = sample.getParent();
-    if (tempParent.getId() != Sample.UNSAVED_ID) {
+    if (tempParent.isSaved()) {
       Sample parent = sampleStore.getSample(tempParent.getId());
       if (parent == null)
         throw new IllegalArgumentException("Parent sample does not exist");
@@ -552,7 +551,7 @@ public class DefaultSampleService implements SampleService, PaginatedDataSource<
       slides += ((SampleTissuePiece) child).getSlidesConsumed();
       parentSlides.setSlides(slides);
       parentSlides.setDiscards(0);
-      if (parentSlides.getId() != SampleImpl.UNSAVED_ID) {
+      if (parentSlides.isSaved()) {
         update(parentSlides);
       }
     }
@@ -596,7 +595,7 @@ public class DefaultSampleService implements SampleService, PaginatedDataSource<
   }
 
   private SampleIdentity findOrCreateIdentity(DetailedSample descendant, SampleIdentity identity) throws IOException, MisoNamingException {
-    if (identity.getId() != SampleImpl.UNSAVED_ID) {
+    if (identity.isSaved()) {
       return (SampleIdentity) sampleStore.getSample(identity.getId());
     } else {
       // If samples are being bulk received for the same new donor, they will all have a null parentId.
@@ -1105,7 +1104,7 @@ public class DefaultSampleService implements SampleService, PaginatedDataSource<
 
   @Override
   public Sample save(Sample sample) throws IOException {
-    if (sample.getId() == SampleImpl.UNSAVED_ID) {
+    if (!sample.isSaved()) {
       return get(create(sample));
     } else {
       update(sample);